backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
aiomysql
python-dotenv
PyJWT
aiohttp